+++ src/wtp_game.h
@@ -1266,4 +1266,5 @@ int getUnitOffenseExtendedTriad(int unitId, int enemyUnitId);
 int getBaseDoctorCount(int baseId);
 int getBaseSpecialistPsych(int baseId);
 bool isFriendlyBaseInRangeHasFacility(int factionId, int x, int y, int range, FacilityId facilityId);
+int getSupportCost(int factionId);